requirements.txt
fastapi>=0.95.0
uvicorn>=0.20.0
python-dotenv>=0.20.0
twilio>=7.0.0
websockets>=10.0
requests>=2.28.0
transformers>=4.30.0
langchain>=0.0.200
chromadb>=0.3.0
torch>=2.0.0
sentence-transformers>=2.2.0
python-multipart
